extract mobile scroller into ambiguous component
